# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Telegram Drive API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local server"
    }
  ],
...
  "paths": {
...
    "/folders": {
      "get": {
        "tags": [
          "Folders"
        ],
        "summary": "Scan and return all Telegram Drive folders",
        "parameters": [
          {
            "in": "header",
            "name": "x-session-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of folders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "parent_id": {
                            "type": "number",
                            "nullable": true
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "data": [
                    {
                      "id": -100123456,
                      "name": "Documents",
                      "parent_id": null
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Folders"
        ],
        "summary": "Create a new folder (Channel)",
        "parameters": [
          {
            "in": "header",
            "name": "x-session-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "parent_id": {
                    "type": "number",
                    "description": "Optional parent folder ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Folder created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "name": {
                          "type": "string"
                        },
                        "parent_id": {
                          "type": "number",
                          "nullable": true
                        }
                      }
                    }
                  }
                },
                "example": {
                  "data": {
                    "id": -100123456,
                    "name": "Photos",
                    "parent_id": null
                  }
                }
              }
            }
          }
        }
      }
    },
    "/folders/{id}": {
      "delete": {
        "tags": [
          "Folders"
        ],
        "summary": "Delete a folder",
        "parameters": [
          {
            "in": "header",
            "name": "x-session-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Folder deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "status": "deleted"
                }
              }
            }
          }
        }
      }
    },
    "/files": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "Get files in a folder",
        "parameters": [
          {
            "in": "header",
            "name": "x-session-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "folder_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of files",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "folder_id": {
                            "type": "number",
                            "nullable": true
                          },
                          "name": {
                            "type": "string"
                          },
                          "size": {
                            "type": "number"
                          },
                          "mime_type": {
                            "type": "string"
                          },
                          "file_ext": {
                            "type": "string"
                          },
                          "created_at": {
                            "type": "string"
                          },
                          "icon_type": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "data": [
                    {
                      "id": 42,
                      "folder_id": -100123456,
                      "name": "report.pdf",
                      "size": 1048576,
                      "mime_type": "application/pdf",
                      "file_ext": "pdf",
                      "created_at": "2026-05-01T12:00:00.000Z",
                      "icon_type": "file"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Files"
        ],
        "summary": "Delete files",
        "parameters": [
          {
            "in": "header",
            "name": "x-session-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message_ids": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    }
                  },
                  "folder_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Files deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "status": "deleted"
                }
              }
            }
          }
        }
      }
    },
    "/files/upload": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Upload multiple files",
        "parameters": [
          {
            "in": "header",
            "name": "x-session-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "folder_id": {
                    "type": "string"
                  },
                  "transfer_id": {
                    "type": "string"
                  },
                  "transfer_ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Files uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "status": "uploaded"
                }
              }
            }
          }
        }
      }
    },
    "/files/move": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Move files between folders",
        "parameters": [
          {
            "in": "header",
            "name": "x-session-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message_ids": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    }
                  },
                  "source_folder_id": {
                    "type": "string"
                  },
                  "target_folder_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Files moved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "status": "moved"
                }
              }
            }
          }
        }
      }
    },
...
  }
}